pyproject.toml
[tool.poetry]
name = "meme-bot-project"
version = "0.1.0"
description = ""
authors = ["RandomguyLel"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
python-dotenv = "^0.15.0"
flask = "^1.1.2"
instagram-private-api = "^1.6.0"
ctx = "^0.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
